Add files via upload
Jupyter notebook project for a data analysis app created using some of the most popular Python libraries and APIs (Pandas | Numpy | Matplotlib | Seaborn  | Tkinter | Pymongo )
A landing window enables the user to select an initial action: selecting available CSV files, importing files from a database or exiting the program.
The Data Preparation window assists users in importing, converting,cleaning and saving data into a MongoDB database.
Finally, the Data manipulation and visualization windows provide a means to view the data in graphic form.
In order to enhance efficiency and assist users in recognizing patterns as well as quickly familiarize themselves with the system, the design of each window has been kept identical with buttons being the only element utilized in different sizes and various activation states.
Buttons are strategically placed to draw user’s attention and visibility is further improved by the flow-structure, which is kept linear , presenting the output of different actions within the same window.UI elements such as buttons are designed to be active, inactive or highlighted according to the state that the application is in at any given moment. 
As an example, the save button is visible only when data manipulation tasks are complete, while the exit button materializes only after the data has been saved.
Colors and texture are an exceptional medium for drawing attention to a certain element.
For this reason, windows have been kept neutral in color while buttons indicating either activation or deactivation have been designed with contrasting colors.
Potential errors such as “files not found” or wrong formatting are identified by the application using exception handling and appropriate messages are then displayed to the user whom is kept informed throughout the process. User needs have been anticipated by setting default names to be used for storage and also files to be selected in the presence or absence of prepared data.